% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_complete_func.R
\name{add_name}
\alias{add_name}
\title{rename address variables
\code{add_name} is used for renaming address variables.}
\usage{
add_name(dat, num)
}
\arguments{
\item{dat}{A data.frame}

\item{num}{Num of address variables}
}
\value{
A data.frame with new name.
}
\description{
rename address variables
\code{add_name} is used for renaming address variables.
}
