% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_distance_func.R
\name{address_distance}
\alias{address_distance}
\title{Address distance}
\usage{
address_distance(add)
}
\arguments{
\item{add}{list of longitude and latitude.}
}
\value{
A list contains address distance
}
\description{
\code{address_distance} is for computing distance between multi sets of address.
}
\examples{
add = list(x = c(120,30),y = c(120,20),z = c(120,10))
address_distance(add)
}
\seealso{
\code{\link{multi_address_distance}}
}
