% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.R
\name{ascat}
\alias{ascat}
\title{Convert number to factor by inner breaks}
\usage{
ascat(x, breaks, labels = NULL, include.lowest = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{breaks}{a numeric vector of two or more unique cut points}

\item{labels}{labels for the levels of the resulting category}

\item{include.lowest}{logical; if \code{TRUE}, include the minimum value in the
first interval. Defaults to \code{TRUE}.}
}
\value{
A \code{factor} is returned
}
\description{
Convert number to factor by inner breaks
}
\examples{
ascat(1:10, 5) # eqaul cut(1:10, c(1, 5, 10), include.lowest = TRUE)
ascat(1:10, 5, c("less than or euqal to 5", "greater than 5"))
}
