% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{cal_treated_start_time}
\alias{cal_treated_start_time}
\title{calculate the time when an individual was first treated}
\usage{
cal_treated_start_time(time, treat)
}
\arguments{
\item{time}{time}

\item{treat}{whether treated at specific time}
}
\description{
calculate the time when an individual was first treated
}
