% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{calpscore_panel}
\alias{calpscore_panel}
\title{calculate propensity score for panel data}
\usage{
calpscore_panel(data, treat, cov, lag = 0L, method = "logit")
}
\arguments{
\item{data}{a data.table labeld as panel data by \code{stxtset()}}

\item{treat}{variable name that indicates whether the unit is treated}

\item{cov}{character vector indicating the set of covarieates}

\item{lag}{an integer vector indicating how to use lags of covarieates}

\item{method}{a string indicating the method used to calcuate propensity
score which will be passed to \code{binomial()} as augument \code{link}.}
}
\description{
calculate propensity score for panel data
}
