% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_archive.R
\name{check_attr}
\alias{check_attr}
\title{Check attributes of data.frame or vetor}
\usage{
check_attr(x, quietly = FALSE)
}
\arguments{
\item{x}{a R object for checking}

\item{quietly}{Bool value, whether output attributes}
}
\description{
\emph{Personal use!}
}
