% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_complete_func.R
\name{complete_address}
\alias{complete_address}
\title{Complete addresses}
\usage{
complete_address(address, key = NULL, map = "baidu")
}
\arguments{
\item{address}{Name of address variable of dat, or a list of addresses.}

\item{key}{Key of map. Apply from \url{https://lbs.amap.com/api/webservice/guide/api/georegeo/} or \url{https://developer.baidu.com/map/android-mobile-apply-key.htm}}

\item{map}{Which map API, 'baidu' and 'ampap' are available.}
}
\value{
A list contains geo information
}
\description{
\code{complete_address} is used for getting complete address from map API.
}
\examples{
data(dict_add_cn)
addr = paste(substr(dict_add_cn[[1]][1],1,2),substr(dict_add_cn[[1]][2],1,2),
substr(dict_add_cn[[1]][3],1,3),substr(dict_add_cn[[1]][4],1,3))
complete_address(addr)
}
\seealso{
\code{\link{multi_complete_address}},\code{\link{complete_address_all}}
}
