% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_complete_func.R
\name{complete_address_all}
\alias{complete_address_all}
\title{Complete address all}
\usage{
complete_address_all(dat, x_list = NULL, ID = NULL, ex_cols = NULL,
  key = NULL, map = NULL, parallel = FALSE)
}
\arguments{
\item{dat}{data.frame contains address variables.}

\item{x_list}{Names of addresses.}

\item{ID}{The name of ID of observations. Default is NULL.}

\item{ex_cols}{A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}

\item{key}{Key of map. Apply from \url{https://lbs.amap.com/api/webservice/guide/api/georegeo/} or \url{https://developer.baidu.com/map/android-mobile-apply-key.htm}}

\item{map}{Which map API, 'baidu' and 'ampap' are available.}

\item{parallel}{Parallel computing option.Default is FALSE.}
}
\value{
A data.frame contains geo information
}
\description{
\code{complete_address_all} is used for multiple address geographic coding of all users.
}
\examples{
data(dict_add_cn)
addr = paste(substr(dict_add_cn[[1]][1],1,2),substr(dict_add_cn[[1]][2],1,2),
substr(dict_add_cn[[1]][3],1,3),substr(dict_add_cn[[1]][4],1,3))
}
\seealso{
\code{\link{complete_address}},\code{\link{multi_complete_address}}
}
