% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_words_func.R
\name{cut_words}
\alias{cut_words}
\title{Cut words}
\usage{
cut_words(string, lowercase = TRUE, engine = NULL, quick_seg = TRUE)
}
\arguments{
\item{string}{A string.}

\item{lowercase}{Whether transform string to lowercase. Default is TRUE.}

\item{engine}{Words segment engine of jiebar packages.}

\item{quick_seg}{Logical,quick segmentation of sentence.Default is TRUE.}
}
\description{
\code{cut_words} is used for cutting sentences as words.
}
\examples{
data(dict_add_cn)
cut_words(paste(substr(dict_add_cn[[1]][1],1,2),substr(dict_add_cn[[1]][2],1,2),
substr(dict_add_cn[[1]][3],1,3)))
}
