% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_archive.R
\name{df2parquet}
\alias{df2parquet}
\title{Write data frame to Parquet file}
\usage{
df2parquet(df, database, table, replace = FALSE, append = FALSE)
}
\arguments{
\item{df}{A data frame of values.}

\item{database}{Database name.}

\item{table}{Table name.}

\item{replace}{logical. If TRUE, overwrite existing file. Default FALSE.}

\item{append}{logical. If TRUE, read existing file, rbind and rewrite. Default FALSE.}
}
\value{
TRUE invisibly on success, NA if file exists and neither replace nor append.
}
\description{
Writes, replaces or appends a data frame to a Parquet file.
File path: \verb{$DATA_ARCHIVE/<database>_<table>.parquet}
}
