% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2sqlite.R
\name{df2sqlite}
\alias{df2sqlite}
\title{Write data frame to database}
\usage{
df2sqlite(df, database, table, keys, replace = FALSE, append = FALSE)
}
\arguments{
\item{df}{A data frame of values (or coercible to data.frame).}

\item{database}{Database name, which will be converted to a database. If
environment variable \code{DATA_ARCHIVE} has been set, then the \code{database} will
be transformed to \verb{$DATA_ARCHIVE/<database>.sqlite}, otherwise, the
\code{database} will be transformed to \verb{$HOME/Data/DBMS/<database>.sqlite.}}

\item{table}{Table name in the database}

\item{keys}{character vector, primary keys of data.frame df}

\item{append}{logical value, whether append \code{df} to the table when it
already exists. default: \code{FALSE}}

\item{reaplace}{logical value, whether replace the \code{table} when it already
exists. default: \code{FALSE}}
}
\description{
Writes, replace of append a data frame to a database table. At
the same time, setting the primary keys of the table.
}
\examples{
\dontrun{
df <- mtcars
df$ID <- seq_along(nrows(df))
df2sqlite(df, database = "test", table = "mtcars", keys = "ID")

df$ID = df$ID + 100
try(df2sqlite(df, "test", "mtcars", "ID", append = TRUE))
df2sqlite(df, "test", "mtcars", "ID", append = TRUE)

df$ID = df$ID + 100
df2sqlite(df, "test", "mtcars", "ID", replace = TRUE)
}
}
