% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{event_study_plot}
\alias{event_study_plot}
\title{draw event study ploy with event study coefficients}
\usage{
event_study_plot(
  coef_result,
  xlabel = "Exposure",
  ylabel = "Coefficient Estimate"
)
}
\arguments{
\item{coef_result}{a data.frame with columns \code{exposure}, \code{coef}, \code{se}
(typically the output of \code{event_study_result()}).}

\item{xlabel}{x-axis label.}

\item{ylabel}{y-axis label.}
}
\description{
draw event study ploy with event study coefficients
}
