% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{execute_box_mod_func}
\alias{execute_box_mod_func}
\title{Execute a Function from a Dynamically Loaded Box Module}
\usage{
execute_box_mod_func(mod, func, args)
}
\arguments{
\item{mod}{A character string specifying the module path or name.}

\item{func}{A character string specifying the function name to retrieve from the module.}

\item{args}{A list of arguments to pass to the function.}
}
\value{
The result of the executed function, or \code{NULL} if an error occurs.
}
\description{
This function dynamically loads a specified module using \code{box}, retrieves a
specific function from it, and executes that function with the provided
arguments.
}
