% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_type.R
\name{expr_type}
\alias{expr_type}
\title{expression type form AdvanceR by Hadley Wickham}
\usage{
expr_type(x)
}
\arguments{
\item{x}{expression}
}
\value{
a string represent the expression type
}
\description{
expression type form AdvanceR by Hadley Wickham
}
