% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{fill}
\alias{fill}
\title{Replacing empty value with latest non-empty value}
\usage{
fill(x, forward = TRUE, maxgap = Inf)
}
\arguments{
\item{x}{an R object to be filled}

\item{forward}{bool value, define filling direction}

\item{maxgap}{Int value, define max fill gap}
}
\description{
based on https://stackoverflow.com/a/13810615
}
\examples{
x = c(NA,NA,'a',NA,NA,NA,NA,NA,NA,NA,NA,'b','c','d',NA,NA,NA,NA,NA,'e') 
fill(x)
fill(x, forward = FALSE)
fill(x, maxgap = 4)
fill(x, maxgap = 5)
}
