% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_lang.R
\name{fmt_lang}
\alias{fmt_lang}
\title{format character vector}
\usage{
fmt_lang(x, lang = "en_US")
}
\arguments{
\item{x}{a character vector}

\item{lang}{a string. The default value is "zh_CN" (Chinese) or
"en_US" (English, default). "zh_CN" keeps CJK ideographs.}
}
\description{
Convert lowercase to uppercase, full-width characters to
half-width, and drop meaningless characters according to language.
}
\examples{
x <- c("\uff22  ", "\uff23\uff10\uff11\uff12,,,\uff13", "cabc=dd")
fmt_lang(x)
}
