% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_scale_transform_function.R
\name{gen_scale_transform_function}
\alias{gen_scale_transform_function}
\title{生成从一个坐标轴到另一个坐标轴的转化函数}
\usage{
gen_scale_transform_function(x, y)
}
\arguments{
\item{x}{两个数字组成的向量，且必须从小到大排列}

\item{y}{两个数字组成的向量，且必须从小到大排列}
}
\value{
返回两个函数构成的列表
}
\description{
建立两个坐标轴之间的映射，方便在 \code{ggplot} 上做双轴图
}
\author{
Liubianshi
}
