% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj.R
\name{get_df_names}
\alias{get_df_names}
\title{Get variable names from data.frame or list}
\usage{
get_df_names(data, ...)
}
\arguments{
\item{data}{a data.frame or list}

\item{...}{symbol, character vector of names}
}
\description{
Get variable names from data.frame or list
}
\examples{
get_df_names(mtcars, mpg, cyl, disp, hp)
get_df_names(mtcars, c("mpg", "cyl", "disp", "hp"))
x <- c("mpg", "cyl", "disp", "hp")
get_df_names(mtcars, x)
x <- c("mpg", "cyl", "disp", "hp2")
try(get_df_names(mtcars, x))
}
