% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getalltable}
\alias{getalltable}
\title{get all tables' basic information from srmd_repo}
\usage{
getalltable()
}
\value{
All tables name and label stored in srdm repo
}
\description{
get all tables' basic information from srmd_repo
}
