% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getallvar}
\alias{getallvar}
\title{get all variables' basic infomation in srmd repo}
\usage{
getallvar()
}
\value{
All variable name and label stored in srdm repo
}
\description{
get all variables' basic infomation in srmd repo
}
