% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getDataSQLite}
\alias{getDataSQLite}
\title{Get data from SQLite Database}
\usage{
getDataSQLite(
  database,
  table,
  varlist = NULL,
  condition = NULL,
  path = NULL,
  and = TRUE,
  limit = NULL
)
}
\arguments{
\item{database}{SQLite database path}

\item{table}{Table name}

\item{varlist}{A character vector consist of column names.}

\item{condition}{A character vector consist of SQL conditions}

\item{and}{Logical value. Way of combine conditions}

\item{limit}{A interger. SQL Limit.}
}
\value{
A data.frame
}
\description{
从 SQLite 数据库中获取数据表
}
