% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{\%//\%}
\alias{\%//\%}
\title{\verb{\%//\%}: replace NULL with default value}
\usage{
x \%//\% then
}
\arguments{
\item{x, }{an object}

\item{then, }{return value when \code{x} is \code{NULL}}
}
\description{
\verb{\%//\%}: replace NULL with default value
}
