% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{ifempty}
\alias{ifempty}
\title{ifempty: replace empty elements with default value}
\usage{
ifempty(x, then, na = NULL)
}
\arguments{
\item{x, }{an object}

\item{then, }{default value for empty x}

\item{na, }{values can be considered as empty value}
}
\description{
ifempty: replace empty elements with default value
}
