% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{ifthen}
\alias{ifthen}
\title{Scalar version of ifelse}
\usage{
ifthen(x, then, otherwise = x, fun = is.null)
}
\arguments{
\item{x}{An object to test.}

\item{then}{Return value when \code{do.call(fun, list(x))} is \code{TRUE}.}

\item{otherwise}{Return value when \code{do.call(fun, list(x))} is \code{FALSE}.
Defaults to \code{x}.}

\item{fun}{A function or function name used as the predicate.
Defaults to \code{is.null}.}
}
\value{
\code{then} or \code{otherwise} depending on \code{fun(x)}.
}
\description{
Returns \code{then} when \code{fun(x)} is \code{TRUE}, otherwise returns \code{otherwise}.
}
