% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{isempty}
\alias{isempty}
\title{'Not Available' / Missing Values / Empty values / NULL}
\usage{
isempty(x, na = NULL)
}
\arguments{
\item{x}{an R object to be tested}

\item{na}{values can be treated as empty, default is NULL}
}
\description{
In the actual data processing process, the appearance of
a null value is either because it is not available, or because the value is
invalid, or because it does not exist. Sometimes it is necessary to make
a strict distinction, but in many cases, it is only necessary to judge
whether a value is Empty value is sufficient.
}
\examples{
x <- c("a", "b", "c", " ", "", NA, "#N/A", "--")
isempty(x)
isempty(x, "#N/A")
isempty(x, c("#N/A", "--"))
y <- c(1, 2, 3, 0, 8, NA, 99, NA, 99)
isempty(y)
isempty(y, 0)
isempty(y, c(0, 99))
}
