% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_lng_to_add.R
\name{lat_lng_to_address}
\alias{lat_lng_to_address}
\title{Transform longitude and latitude to address}
\usage{
lat_lng_to_address(lng = NULL, lat = NULL, map = "baidu",
  key = NULL, coordtype = "wgs84ll")
}
\arguments{
\item{lng}{A list of Longitudes.}

\item{lat}{A list of Latitudes.}

\item{map}{Which map API, 'baidu' and 'amap' are available.}

\item{key}{Key of map. Apply from \url{https://lbs.amap.com/api/webservice/guide/api/georegeo/}
or \url{https://developer.baidu.com/map/android-mobile-apply-key.htm}}

\item{coordtype}{Coordtype of longitude and latitude. Default is wgs8411.}
}
\value{
A list of formmated address.
}
\description{
\code{lat_lng_to_address} is used for transforming longitude and latitude to address.
}
\examples{
lat_lng_to_address(lat = 31.231078,lng = 121.699373,key =  NULL)
}
