% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbs-package.R
\docType{package}
\name{lbs-package}
\alias{lbs-package}
\alias{lbs}
\title{Personal use fucntion}
\description{
R functions for pernsonal use
}
\author{
\strong{Maintainer}: Wei Luo \email{liu.bian.shi@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
