% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbs_theme_minimal.R
\name{lbs_theme_miminal}
\alias{lbs_theme_miminal}
\title{: ggplot theme, based on \code{cowplot} 的 \code{theme_minimal_grid}}
\usage{
lbs_theme_miminal(
  axis.title.x = element_blank(),
  legend.position = "top",
  legend.justification = "center",
  legend.box = "horizontal",
  dpi = 600L,
  fontfamily = "Noto Serif CJK SC",
  fontfile = "NotoSerifCJK-Regular.ttc"
)
}
\description{
Custom theme for ggplot: minimal
}
