% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{list_tables}
\alias{list_tables}
\alias{getalltable}
\title{List all archived tables}
\usage{
list_tables()

getalltable()
}
\value{
A data.table with columns database, table, keys, description.
}
\description{
\emph{Personal use!} Return all table names and descriptions in the
metadata index.
}
