% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{list_variables}
\alias{list_variables}
\alias{getallvar}
\title{List all archived variables}
\usage{
list_variables()

getallvar()
}
\value{
A data.table with columns database, table, variable, label.
}
\description{
\emph{Personal use!} Return all variable names and labels in the
metadata index.
}
