% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_distance_func.R
\name{multi_address_distance}
\alias{multi_address_distance}
\title{Multiple address distance}
\usage{
multi_address_distance(dat, lat = "lat", lng = "lng", ID = NULL,
  parallel = FALSE)
}
\arguments{
\item{dat}{data.frame contains address variables.}

\item{lat}{Name of Latitude variable of dat, or a list of Latitudes.}

\item{lng}{Name of Longitude variable of dat, or a list of Longitudes.}

\item{ID}{The name of ID of observations. Default is NULL.}

\item{parallel}{Parallel computing option.Default is FALSE.}
}
\value{
A data.frame contains address distance.
}
\description{
\code{multi_address_distance} is for computing  distance between multi sets of addresses.
}
\examples{
add = rbind(add1 = c(x = c(120,30),y = c(120,20),z = c(120,10)),
                add2 = c(x = c(130,30),y = c(140,30),z = c(120,30)))
multi_address_distance(dat = add,lat = "2",lng ="1")
}
\seealso{
\code{\link{address_distance}}
}
