% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_complete_func.R
\name{multi_complete_address}
\alias{multi_complete_address}
\title{Multiple address geographic coding
\code{multi_complete_address} is used for multiple address geographic coding.}
\usage{
multi_complete_address(dat, x_list = NULL, ex_cols = NULL,
  map = NULL, key = NULL)
}
\arguments{
\item{dat}{data.frame contains address variables.}

\item{x_list}{Names of addresses.}

\item{ex_cols}{A list of excluded variables. Regular expressions can also be used to match variable names. Default is NULL.}

\item{map}{Which map API, 'baidu' and 'amap' are available.}

\item{key}{Key of map. Apply from \url{https://lbs.amap.com/api/webservice/guide/api/georegeo/} or \url{https://developer.baidu.com/map/android-mobile-apply-key.htm}}
}
\value{
A data.frame
}
\description{
Multiple address geographic coding
\code{multi_complete_address} is used for multiple address geographic coding.
}
\examples{
data(dict_add_cn)
addr = paste(substr(dict_add_cn[[1]][1],1,2),substr(dict_add_cn[[1]][2],1,2),
substr(dict_add_cn[[1]][3],1,3),substr(dict_add_cn[[1]][4],1,3))
multi_complete_address(addr)
}
\seealso{
\code{\link{complete_address}},\code{\link{complete_address_all}}
}
