% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{print_flextable}
\alias{print_flextable}
\title{print auxiliary table for flextable object in order to use pandoc-crossref}
\usage{
print_flextable(ft, caption = NULL, label = NULL)
}
\arguments{
\item{ft}{a flextable object}

\item{caption}{a string for table caption}

\item{label}{a valid name for pondoc-crossref}
}
\description{
Used in Rmarkdown like file with \code{results = "asis"}
}
