% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push-pop-shift-unshift.R
\name{push}
\alias{push}
\title{perl-style push}
\usage{
push(x, ...)
}
\arguments{
\item{x}{a vector or list in the caller's frame; modified in place.}

\item{...}{values (optionally named) to append.}
}
\description{
perl-style push
}
