% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{setgraphelements}
\alias{setgraphelements}
\title{set default graph elements}
\usage{
setgraphelements(
  dpi = 600L,
  linetype = "solid",
  linecolor = "black",
  fontcolor = "black",
  fontfamily = "Noto Serif CJK SC"
)
}
\arguments{
\item{dpi}{integer; output DPI used to scale font sizes.}

\item{linetype}{default line type used by built-in element helpers.}

\item{linecolor}{default line/grid colour.}

\item{fontcolor}{default text colour.}

\item{fontfamily}{default text family.}
}
\description{
set default graph elements
}
