% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push-pop-shift-unshift.R
\name{shift}
\alias{shift}
\title{perl-style shift}
\usage{
shift(x, drop = TRUE)
}
\arguments{
\item{x}{a vector or list in the caller's frame; modified in place.}

\item{drop}{logical; if \code{TRUE}, unwrap a single-element list result.}
}
\description{
perl-style shift
}
