% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stenstring}
\alias{stenstring}
\title{Trans numeric vector to character with specific format}
\usage{
stenstring(x, fmt = "\%d")
}
\arguments{
\item{x}{a numeric vector}

\item{fmt}{a character vector of format strings}
}
\description{
Trans numeric vector to character with specific format
}
