% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stlabel}
\alias{stlabel}
\title{Set attributes for variables in data.frame}
\usage{
stlabel(df, var, attr, type = "label")
}
\arguments{
\item{df}{data.table}

\item{var}{one bare variable name or character vector of variable names}

\item{attr}{numeric or character vector, length must equal \code{var}}

\item{type}{name of the attribute to set (default \code{"label"})}
}
\description{
Set attributes for variables in data.table
}
