% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{stlag.data.frame}
\alias{stlag.data.frame}
\title{stlag for data.frame}
\usage{
\method{stlag}{data.frame}(df, varlist = NULL, time = NULL, by = NULL, n = 1L, mode = NULL)
}
\description{
stlag for data.frame
}
\examples{
ts <- data.table(time = 2001:2009, x = 1:9, y = sample(letters, 9))
stlag(ts, x, time = "time", mode = "list")
stlag(ts, x, time)
stlag(ts, y, time, n = -1L)
xt <- data.frame(
    id = rep(c("a", "b"), each = 5),
    time = rep(2001:2005, 2),
    x = 1:10,
    y = sample(letters, 10, replace = TRUE),
    z = sample(LETTERS, 10, replace = TRUE)
)
stxtset(xt, id, time)
stlag(xt, x, n = -1L, mode = "list")
stlag(setDT(xt), x)
xt

}
