% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{stlag.default}
\alias{stlag.default}
\title{Calculate lagged value}
\usage{
\method{stlag}{default}(x, time, n = 1L)
}
\arguments{
\item{x}{a vector}

\item{time}{a interger vector, length must equal to x}

\item{n}{an integer}
}
\description{
Calculate n lagged value of x relate to time
}
\examples{
time = rep(2001:2006)
x = sample(6)
stlag(x, time)
}
