% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{sttscheck}
\alias{sttscheck}
\title{check whether a data.frame is a time-series table}
\usage{
sttscheck(df)
}
\arguments{
\item{df}{data.frame object}
}
\description{
check whether a data.frame is a time-series table
}
