% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{sttsset}
\alias{sttsset}
\title{set data.frame as time-series data}
\usage{
sttsset(df, time)
}
\arguments{
\item{df}{a data.table object}

\item{time}{variable name as time}
}
\description{
Mark \code{data.table} as Panel data, and set \code{id} and \code{time}
}
