% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stxtcheck}
\alias{stxtcheck}
\title{check whether a data.table is a panel table}
\usage{
stxtcheck(dt)
}
\arguments{
\item{dt}{data.table object}
}
\description{
check whether a data.table is a panel table
}
