% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stxtlag}
\alias{stxtlag}
\title{Stimulate panel time operator in stata}
\usage{
stxtlag(panel.dt, varlist, n = 1L, new.dt = FALSE)
}
\arguments{
\item{panel.dt}{an data.table object with a TRUE attribute "panel",
which is setted through \code{stxtset()}}

\item{varlist}{variable names needed to calculate lagged value}

\item{n}{an interger, determine lag direction and interval}

\item{new.df}{a logic value, determine whether to generate a new data.table.
default is \code{FALSE}}
}
\description{
stata-style panel time operator.
}
\examples{
library(data.table)
dt <- data.table(
    id = rep(c("a", "b", "c", "d"), each = 4),
    time = rep(2001:2004, 4),
    x = sample(16)
)
stxtset(dt, id, time)
(stxtlag(dt, x))
}
