% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{stxtset}
\alias{stxtset}
\title{set data.frame as panel data}
\usage{
stxtset(df, id, time)
}
\arguments{
\item{df}{a data.table object}

\item{id}{variable name as panel id}

\item{time}{variable name as panel time}
}
\description{
Mark \code{data.table} as Panel data, and set \code{id} and \code{time}
}
