% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{trans_scale_factory}
\alias{trans_scale_factory}
\title{generate scale transform function pair}
\usage{
trans_scale_factory(x, y)
}
\arguments{
\item{x}{a numeric vector with two elements}

\item{y}{a numeric vector with two elements}
}
\value{
a list contain two reciprocal functions
}
\description{
generate scale transform function pair
}
