% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push-pop-shift-unshift.R
\name{unshift}
\alias{unshift}
\title{perl-style unshift}
\usage{
unshift(x, ...)
}
\arguments{
\item{x}{a vector or list in the caller's frame; modified in place.}

\item{...}{values (optionally named) to prepend.}
}
\description{
perl-style unshift
}
