% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2sqlite.R
\name{df_srdm}
\alias{df_srdm}
\title{Prepare data.frame information for SRDM}
\usage{
df_srdm(
  df,
  database,
  table,
  replace = FALSE,
  append = FALSE,
  wirte_repo = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{data.frame for archive}
}
\description{
\emph{Personal use!}. Check whether the data frame meets the
requirements of SRDM, and output the basic information of the data to
standard output
}
