% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{event_study_plot}
\alias{event_study_plot}
\title{draw event study ploy with event study coefficients}
\usage{
event_study_plot(
  coef_result,
  xlabel = "Exposure",
  ylabel = "Coefficient Estimate"
)
}
\description{
draw event study ploy with event study coefficients
}
