% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{event_study_result}
\alias{event_study_result}
\title{execute pre-trend test for did analysis}
\usage{
event_study_result(
  data,
  id,
  time,
  output,
  treat_status = NULL,
  group = NULL,
  never_treat = NULL,
  covs = NULL,
  drop_exposure_list = -1
)
}
\description{
execute pre-trend test for did analysis
}
