% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_lang.R
\name{fmt_lang}
\alias{fmt_lang}
\title{format character vector}
\usage{
fmt_lang(x, lang = "en_US")
}
\arguments{
\item{x}{a character vector}

\item{lang}{a string. The default value is "zh_CN"}
}
\description{
将小写字母转化为大写字母，全角符号转化为半角符号，并根据语言剔除无意义字符
}
\examples{
x <- c('Ｂ  ', 'Ｃ０１２,,,３', 'cabc=dd')
fmt_lang(x)
y <- c(x, "中文３４５６Ｚａｂ")
fmt_lang(y)
fmt_lang(y, lang = "zh_CN")
}
