% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_statistic_table.R
\name{gen_statistic_table}
\alias{gen_statistic_table}
\title{Generate Statistic Result table}
\usage{
gen_statistic_table(
  stat,
  p,
  se = NULL,
  t = NULL,
  format = c(stat = "3", se = "[3]"),
  star = list(cut = c(0.1, 0.05, 0.01), symbol = c("*", "**", "***")),
  outfmt = "md"
)
}
\arguments{
\item{stat}{statistic data.frame}

\item{p}{p Value data.frame}

\item{se}{standard error data.frame}

\item{t}{t statistic  data.frame}

\item{star}{a list contain cut and symbol}

\item{foramt}{a list contain format information}
}
\description{
Gen Academic table
}
