% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regformula.R
\name{genformula}
\alias{genformula}
\title{gen regress formula}
\usage{
genformula(
  varlist,
  method = NULL,
  fe = NULL,
  cluster = NULL,
  en = NULL,
  iv = NULL,
  ...
)
}
\arguments{
\item{dep:}{dependent variable, require}

\item{indep_ex:}{exogeneous independent variables, require}

\item{fe:}{fixed effect, optional}

\item{cluster:}{cluster varaible, optional}

\item{en:}{endogenous independent variables, optional}

\item{iv:}{Instrumental variables, must longer than \code{en}}

\item{method:}{Formula usage scenario, default is felm}
}
\description{
gen regress formula
}
