% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getdatainfo}
\alias{getdatainfo}
\title{Get infomation form data.repo}
\usage{
getdatainfo(database, table, var = NULL)
}
\arguments{
\item{database}{SQLite database path}

\item{table}{Table name}
}
\description{
从数据管理库获取数据信息
}
