% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getdatainfo}
\alias{getdatainfo}
\title{Get infomation form data.repo}
\usage{
getdatainfo(database, table, var = NULL)
}
\arguments{
\item{database}{SQLite database path}

\item{table}{Table name}

\item{var}{A character vector consist of column names.}
}
\value{
return a data.frame contain data info
}
\description{
get data information from srdm repo
}
