% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isna.R
\name{ifthen}
\alias{ifthen}
\title{scalar version of ifelse}
\usage{
ifthen(x, then, otherwise = x, fun = is.null)
}
\arguments{
\item{x, }{an object}

\item{then, }{return value when \code{do.call(fun, list(x))} is \code{TRUE}}

\item{otherwise, }{return value when \code{do.call(fun, list(x))} is \code{FALSE}, default value is x}

\item{fun, }{a function or function name, used to judge, default value is \code{is.null}}
}
\description{
scalar version of ifelse
}
