% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbs_theme_minimal.R
\name{lbs_theme_minimal}
\alias{lbs_theme_minimal}
\title{: ggplot theme, based on \code{cowplot} 的 \code{theme_minimal_grid}}
\usage{
lbs_theme_minimal(
  legend.position = "top",
  dpi = 600L,
  fontfamily = getOption("font.serif")
)
}
\description{
Custom theme for ggplot: minimal
}
