% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{nearest_match}
\alias{nearest_match}
\title{nearest match with distance}
\usage{
nearest_match(
  data,
  treat,
  distance,
  discard = "both",
  caliper = NULL,
  std.caliper = FALSE,
  replace = FALSE,
  ...
)
}
\description{
nearest match with distance
}
