% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{psm}
\alias{psm}
\title{execute propensity match and return match result object with match table}
\usage{
psm(
  sample,
  id,
  treat,
  pscore,
  ...,
  match_fun = nearest_match,
  result_handle_fun = NULL
)
}
\arguments{
\item{sample}{a data.table object, one line per individual}

\item{id}{name of a column used to identify individual}

\item{treat}{name of a column used to identify treated status}

\item{pscore}{name of a column for propensity score}

\item{args}{a list contain element to be passed to \code{MatchIt::matchid}}
}
\description{
execute propensity match and return match result object with match table
}
