% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbs-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{:=}
\alias{setattr}
\alias{setDT}
\alias{data.table}
\alias{as.data.table}
\alias{!!}
\alias{!!!}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:assign]{:=}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{data.table}}, \code{\link[data.table]{setDT}}, \code{\link[data.table]{setattr}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:nse-force]{!!}}, \code{\link[rlang:nse-force]{!!!}}}
}}

