% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{setgraphelements}
\alias{setgraphelements}
\title{set default graph elements}
\usage{
setgraphelements(
  dpi = 600L,
  linetype = "solid",
  linecolor = "black",
  fontcolor = "black",
  fontfamily = "Noto Serif CJK SC"
)
}
\description{
set default graph elements
}
