% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stdes}
\alias{stdes}
\title{des variables}
\usage{
stdes(df)
}
\arguments{
\item{df}{\code{data.frame}}
}
\description{
Describe the basic information of variables in stata-style
}
\examples{
df <- data.frame(a = 1:3, b = 2:4)
des(df)

attr(df$a, "label") = "A"
attr(df$b, "label") = "B"
des(df)
}
