% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stformat}
\alias{stformat}
\title{format number in a reasonable way}
\usage{
stformat(
  x,
  digits = 3L,
  nsmall = 3L,
  width = NULL,
  big.mark = ",",
  na.replace = ""
)
}
\arguments{
\item{x, }{numeric vector}

\item{digits, }{how many significant digits are to be used for
numeric and complex \code{x}.  The default, \code{NULL}, uses
\code{getOption("digits")}.}

\item{nsmall, }{the minimum number of digits to the right of the
decimal point in formatting real/complex numbers in
non-scientific formats.}

\item{na.replace, }{used to replace \code{NA}}

\item{width:}{\code{default} method: the \emph{minimum} field width or
\code{NULL} or \code{0} for no restriction.}
}
\description{
format number in a reasonable way.
}
\examples{
stformat(10 ^ seq(-10, 10) + runif(21))
stformat(10 ^ seq(-10, 10) + runif(21), nsmall = 2)
stformat(10 ^ seq(-10, 10) + runif(21), digits = 4)
stformat(10 ^ seq(-10, 10) + runif(21), width = 7)
stformat(10 ^ seq(-10, 10) + runif(21), width = 11)
stformat(c(NA, 10 ^ seq(-5, 5)) + runif(12))
}
