% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stlabel}
\alias{stlabel}
\title{Set attributes for variables in data.table}
\usage{
stlabel(df, variable, attribute, type = "label")
}
\arguments{
\item{df}{data.table}

\item{variable}{one bare variabel name or character vector of variable names}

\item{attributes}{numeric or character vector, length must equal variables}
}
\description{
Set attributes for variables in data.table
}
