% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stlag}
\alias{stlag}
\title{Calculate lagged value}
\usage{
stlag(x, time, n = 1L)
}
\arguments{
\item{x}{a vector}

\item{time}{a interger vector, length must equal to x}

\item{n}{an integer}
}
\description{
Calculate n lagged value of x relate to time
}
