% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stsum}
\alias{stsum}
\title{stata-style sumarisze}
\usage{
stsum(object, ...)
}
\arguments{
\item{object}{numerical vector of data.frame. If the object is a data.frame,
\code{variable} is required.}

\item{viaralbe}{variable name of an character vector consists of variable names}

\item{label}{\code{NULL}, \code{TRUE} or an character vector consists of variable labels.
The length of labels must be equal to the length of variables unless the
label is \code{NULL} or \code{TREU}}
}
\description{
generate a stata-style sumarisze table
}
