% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{stxtcheck}
\alias{stxtcheck}
\title{check whether a data.frame is a panel table --------------------------------}
\usage{
stxtcheck(df)
}
\arguments{
\item{df}{data.frame object}
}
\description{
check whether a data.frame is a panel table --------------------------------
}
