% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_related.R
\name{stxtset}
\alias{stxtset}
\title{set data.table as panel data}
\usage{
stxtset(data, id, time)
}
\arguments{
\item{data}{a data.table object}

\item{id}{variable name as panel id}

\item{time}{variable name as panel time}
}
\description{
Mark \code{data.table} as Panel data, and set \code{id} and \code{time}
}
